+++ src/__tests__/posthog-core.ts
@@ -684,6 +684,21 @@ describe('posthog core', () => {
             expect(posthog.capture).toHaveBeenCalledWith('$pageleave')
         })
 
+        it('captures $pageleave when capture_pageview is set to history_change', () => {
+            const posthog = posthogWith(
+                {
+                    capture_pageview: 'history_change',
+                    capture_pageleave: 'if_capture_pageview',
+                    request_batching: true,
+                },
+                { capture: jest.fn() }
+            )
+
+            posthog._handle_unload()
+
+            expect(posthog.capture).toHaveBeenCalledWith('$pageleave')
+        })
+
         it('does not capture $pageleave when capture_pageview=false and capture_pageleave=if_capture_pageview', () => {
             const posthog = posthogWith(
                 {
@@ -734,6 +749,20 @@ describe('posthog core', () => {
                 const posthog = posthogWith(
                     {
                         capture_pageview: true,
+                        capture_pageleave: 'if_capture_pageview',
+                        request_batching: false,
+                    },
+                    { capture: jest.fn() }
+                )
+                posthog._handle_unload()
+
+                expect(posthog.capture).toHaveBeenCalledWith('$pageleave', null, { transport: 'sendBeacon' })
+            })
+
+            it('captures $pageleave when capture_pageview is set to history_change', () => {
+                const posthog = posthogWith(
+                    {
+                        capture_pageview: 'history_change',
                         capture_pageleave: 'if_capture_pageview',
                         request_batching: false,
                     },