+++ src/extensions/surveys.tsx
@@ -110,7 +110,13 @@ export function getNextSurveyStep(
         if (question.type === SurveyQuestionType.SingleChoice) {
             // :KLUDGE: for now, look up the choiceIndex based on the response
             // TODO: once QuestionTypes.MultipleChoiceQuestion is refactored, pass the selected choiceIndex into this method
-            const selectedChoiceIndex = question.choices.indexOf(`${response}`)
+            let selectedChoiceIndex = question.choices.indexOf(`${response}`)
+
+            if (selectedChoiceIndex === -1 && question.hasOpenChoice) {
+                // if the response is not found in the choices, it must be the open choice,
+                // which is always the last choice
+                selectedChoiceIndex = question.choices.length - 1
+            }
 
             if (question.branching?.responseValues?.hasOwnProperty(selectedChoiceIndex)) {
                 const nextStep = question.branching.responseValues[selectedChoiceIndex]
@@ -240,10 +246,16 @@ export class SurveyManager {
         // Ensure widget container exists if it doesn't
         const shadow = retrieveWidgetShadow(survey, this._posthog)
         const stylesheetContent = style(survey.appearance)
-        const stylesheet = prepareStylesheet(document, stylesheetContent, this._posthog)
+        const WIDGET_STYLE_ID = 'ph-data-style-id'
 
-        if (stylesheet) {
-            shadow.appendChild(stylesheet)
+        // Check if the stylesheet already exists
+        if (!shadow.querySelector(`style[${WIDGET_STYLE_ID}="${WIDGET_STYLE_ID}"]`)) {
+            const stylesheet = prepareStylesheet(document, stylesheetContent, this._posthog)
+
+            if (stylesheet) {
+                stylesheet.setAttribute(WIDGET_STYLE_ID, WIDGET_STYLE_ID) // Add identifier
+                shadow.appendChild(stylesheet)
+            }
         }
 
         Preact.render(